+++ lib/Metadata.php
@@ -41,7 +41,7 @@ class Metadata
     /**
      * @var array
      */
-    private $fieldMappings;
+    private $fieldMappings = [];
 
     /**
      * Add a field mapping for field with given name, for example.